don't let disasm go out of bounds; fast scroll
