+++ src/debugger/stacktrace.c
@@ -6,11 +6,8 @@
 
 #include "map_parser.h"
 #include "symtable.h"
-#include "segment_symbols.h"
 #include "stacktrace.h"
 #include "disasm.h"
-
-extern void __osCleanupThread();
 
 #if defined(DEBUG_EXPORT_SYMBOLS) && defined(DEBUG_FULL_STACK_TRACE)
 
@@ -22,20 +19,6 @@ u32 stackTraceGenerated = FALSE;
 
 static u8 is_top_of_stack(u32 ra) {
     return (ra == ((u32)__osCleanupThread));
-}
-
-static u8 is_text_addr(u32 addr) {
-    if ((addr >= (u32)_mainSegmentStart) && (addr <= (u32)_mainSegmentTextEnd)) {
-        return TRUE;
-    }
-    else if ((addr >= (u32)_engineSegmentStart) && (addr <= (u32)_engineSegmentTextEnd)) {
-        return TRUE;
-    }
-    else if ((addr >= (u32)_goddardSegmentStart) && (addr <= (u32)_goddardSegmentTextEnd)) {
-        return TRUE;
-    }
-
-    return FALSE;
 }
 
 static void add_entry_to_stack(u32 addr, u32 ra, symtable_info_t *info) {
