+++ src/debugger/symtable.c
@@ -35,9 +35,9 @@ static u32 headless_pi_status(void) {
 
 void map_parser_dma(void *dst, void *src, size_t size) {
     osWritebackDCacheAll();
-    osInvalICache(dst, size);
     headless_dma((u32)src, dst, size);
     while (headless_pi_status() & PI_STATUS_IO_BUSY);
+    osInvalICache(dst, size);
 }
 
 /** 
@@ -143,7 +143,7 @@ char *symt_string(symtable_header_t *symt, int sidx, int slen, char *buf, int si
     map_parser_dma(
         func, 
         (uintptr_t *)(SYMT_ROM + symt->strtab_off + sidx),
-        nbytes
+        size
     );
     func[nbytes] = 0;
